pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "i4g"
dynamic = ["version"]
description = "Scratchpad for prototype utilities."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "altair",
    "datasets",
    "faiss-cpu",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "langchain-chroma",
    "langchain-community",
    "langchain-ollama",
    "langchain",
    "ollama",
    "paddleocr[all]",
    "Pillow",
    "fastapi",
    "huggingface-hub>=0.34.0,<1.0",
    "pytesseract",
    "python-docx",
    "regex",
    "requests",
    "rich",
    "streamlit>=1.51.0,<2",
    "tqdm",
    "trio",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-anyio",
    "pip-tools",
    "pre-commit",
]

[project.scripts]
i4g-admin = "i4g.cli.admin:main"
run-dataflow = "prefect_gcp.beam.pipeline:main"

[tool.hatch.version]
path = "VERSION.txt"
pattern = "(?P<version>.+)"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (slower)",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
